Adding parse function
Parse function is working basically as designed, but broke the view
output in doing so.  This is fine for two reasons, the view not being
remotely final and more parsing will be done.

Academically this currently fails with a 500 error, the error resulting
from a flask/wsgi error indicating: "The view function did not return a
valid response. The return type must be a string, tuple, Response
instance, or WSGI callable, but it was a set."

Tried a few different approaches, nothing worked, for now this is
further than before.